fix: refactored fixtures
