+++ modelon/impact/client/entities/workspace.py
@@ -605,7 +605,7 @@ class Workspace:
                 f'No default project exists for the workspace {self._workspace_id}!'
             )
         resp = self._sal.project.project_get(
-            self._workspace_definition.default_project_id
+            self._workspace_definition.default_project_id, vcs_info=True
         )
         return Project(
             resp["id"],
