+++ tests/impact/client/entities/test_model.py
@@ -5,7 +5,6 @@ from tests.impact.client.helpers import (
     create_model_exe_operation,
     IDs,
 )
-from tests.impact.client.fixtures import *
 
 
 class TestModel:
