+++ tests/impact/client/fixtures/jupyterhub.py
@@ -1,7 +1,12 @@
 import pytest
 import requests.exceptions
 
-from tests.impact.client.fixtures import *
+
+from tests.impact.client.helpers import (
+    with_json_route,
+    with_text_route,
+    with_exception,
+)
 
 
 @pytest.fixture
