+++ tests/impact/client/operations/test_experiment.py
@@ -2,7 +2,6 @@ import pytest
 from modelon.impact.client import exceptions
 from tests.impact.client.helpers import create_experiment_entity, IDs
 from modelon.impact.client.operations.base import Status
-from tests.impact.client.fixtures import *
 
 
 class TestExperimentOperation:
