+++ tests/impact/client/operations/test_model_exe.py
@@ -7,7 +7,6 @@ from tests.impact.client.helpers import (
     create_model_exe_operation,
     IDs,
 )
-from tests.impact.client.fixtures import *
 
 
 class TestModelExecutableOperation:
