+++ tests/impact/client/sal/test_custom_function_service.py
@@ -1,7 +1,6 @@
 from modelon.impact.client.sal.uri import URI
 import modelon.impact.client.sal.service
 from tests.impact.client.helpers import IDs
-from tests.impact.client.fixtures import *
 
 
 class TestCustomFunctionService:
