+++ tests/impact/client/sal/test_model_exe_service.py
@@ -1,7 +1,6 @@
 import modelon.impact.client.sal.service
 from modelon.impact.client.sal.uri import URI
 from tests.impact.client.helpers import IDs
-from tests.impact.client.fixtures import *
 
 
 class TestModelExecutbleService:
