+++ tests/impact/client/sal/test_project_service.py
@@ -3,7 +3,6 @@ from modelon.impact.client.sal.uri import URI
 import modelon.impact.client.sal.service
 from tests.impact.client.helpers import IDs
 from tests.files.paths import SINGLE_FILE_LIBRARY_PATH
-from tests.impact.client.fixtures import *
 
 
 class TestProjectService:
