+++ tests/impact/client/sal/test_services.py
@@ -3,7 +3,6 @@ import modelon.impact.client.sal.service
 from modelon.impact.client.sal.uri import URI
 import modelon.impact.client.sal.exceptions
 from tests.impact.client.helpers import IDs, get_test_workspace_definition
-from tests.impact.client.fixtures import *
 
 
 class TestService:
